pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-financial-advisor"
version = "0.1.0"
description = "Production-grade AI Financial Advisor with LangChain agents and RAG"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # OpenAI
    "openai>=1.30.0",

    # Vector store
    "faiss-cpu>=1.8.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Config & validation
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.3.0",

    # Logging
    "structlog>=24.1.0",
    "rich>=13.7.0",

    # Auth (JWT)
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # HTTP client
    "httpx>=0.27.0",

    # Utilities
    "tenacity>=8.3.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "anyio>=4.4.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "UP", "B", "C4", "PTH"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
